move code to helper
